backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
notion-client==2.2.1
feedparser==6.0.10
requests==2.31.0
trafilatura==1.6.4
readability-lxml==0.8.1
tenacity==8.2.3
pandas==2.1.3
openpyxl==3.1.2
pyyaml==6.0.1
loguru==0.7.2
langdetect==1.0.9
tiktoken==0.5.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
